packages/worker/requirements.txt
bidict==0.21.2
click==7.1.2
ffmpeg-python==0.2.0
mysql-connector-python==8.0.23
future==0.18.2
h11==0.12.0
httptools==0.1.1
netifaces==0.10.6
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
SQLAlchemy==1.3.23
uvicorn==0.13.3
uvloop==0.15.0
watchgod==0.6
websockets==8.1
